refactor(nic/amdsmi)[ROCM-27257]: replace lspci NIC stack
Replace AINIC stack with unified transport and vendor registry
NIC support relied on lspci parsing and a sprawl of no-drm device/switch classes duplicated
across the C++ core, the rocm_smi compat layer, the CLI, and Python.

Replace it with a layered, testable AI-NIC subsystem built on a transport abstraction
(netlink preferred, ioctl fallback) and a vendor plugin registry.

It solves: AILITOOLS-294, ROCM-27257

+++ projects/amdsmi/include/amd_smi/impl/nic/amdsmi_unified/interface/smi_nic_interface.h
@@ -166,7 +166,7 @@ typedef struct {
   char type[SMI_NIC_MAX_STRING_LENGTH];
   char flavour[SMI_NIC_MAX_STRING_LENGTH];
   char netdev[SMI_NIC_MAX_STRING_LENGTH];
-  uint8_t ifindex;
+  uint32_t ifindex;
   char mac_address[SMI_NIC_MAX_STRING_LENGTH];
   uint8_t carrier;
   uint16_t mtu;
